Added protocol for message transfer
First, the client will send it's fixed length header containing the message length and it's name. It will then send the message. This is impleneted in the MessageProtocol class in the ChatRoomHelpers module.

When the client first connects, they will send their name to the server who will then assign that name to the connection. The name is got from a popup on the client side.