Client and Sever are now prompted to input host/port
Can now choose what port server listens on.
The server will be accessible to all computers on the network. Client and server no longer crash when the server shutsdown first

Client is also prompted to input host/port of the server. This was done with the use of the new ClientSetUp tkinter component in the helpers module